Add support for floating point numbers:
Numeric expressions/values can now support both int and float types. There are
subtle issues with using `Decimal` to represent these types, primarily in how
binary integer operations (such as and/or/xor) work. Because these are normally
done on integers, but these numeric values can work on floats as well, I
decided it would be best to convert the numeric types to ints, do the binary
operation, then convert it back to a numeric (`Decimal`). This should be
documented once I add docs for the operators.

+++ cicada/eval/constexpr_visitor.py
@@ -1,6 +1,7 @@
 from collections import ChainMap
 from collections.abc import Iterator
 from contextlib import contextmanager
+from decimal import Decimal
 from typing import cast
 
 from cicada.api.domain.triggers import Trigger
@@ -157,19 +158,19 @@ class ConstexprEvalVisitor(NodeVisitor[Value]):
 
             if node.oper == BinaryOperator.DIVIDE:
                 # TODO: move this to semantic layer
-                return NumericValue(int(lhs.value / rhs.value))
+                return NumericValue(lhs.value / rhs.value)
 
             if node.oper == BinaryOperator.MODULUS:
                 return NumericValue(lhs.value % rhs.value)
 
             if node.oper == BinaryOperator.AND:
-                return NumericValue(lhs.value & rhs.value)
+                return NumericValue(Decimal(int(lhs.value) & int(rhs.value)))
 
             if node.oper == BinaryOperator.OR:
-                return NumericValue(lhs.value | rhs.value)
+                return NumericValue(Decimal(int(lhs.value) | int(rhs.value)))
 
             if node.oper == BinaryOperator.XOR:
-                return NumericValue(lhs.value ^ rhs.value)
+                return NumericValue(Decimal(int(lhs.value) ^ int(rhs.value)))
 
             if node.oper == BinaryOperator.LESS_THAN:
                 return BooleanValue(lhs.value < rhs.value)
